feat: extracting widgets from wallet screen
