+++ lib/app_widget.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:projeto_criptos/src/presenter/screens/wallet/wallet_page.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ProviderScope(
+      child: MaterialApp(
+        title: 'Flutter Demo',
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
+        ),
+        home: const WalletScreen(),
+        debugShowCheckedModeBanner: false,
+      ),
+    );
+  }
+}
