--- lib/src/presenter/provider/user_provider.dart
@@ -1,5 +1,0 @@
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:projeto_criptos/src/domain/entities/user_model.dart';
-
-var userProvider = StateProvider((ref) => UserModel());
-
