+++ lib/src/presenter/provider/visible_provider.dart
@@ -1,3 +1,3 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-var visibleProvider = StateProvider((ref) => true);
+final visibleProvider = StateProvider((ref) => true);
