+++ lib/utils/currency_formater.dart
@@ -1,6 +1,6 @@
 import 'package:intl/intl.dart';
 
-NumberFormat formatter = NumberFormat.currency(
+NumberFormat realFormatter = NumberFormat.currency(
   locale: "pt_BR",
   symbol: "R\$"
 );
